support servers like flask by writing modules on import

+++ forkserver/lib/checkpoint.py
@@ -1,5 +1,7 @@
 import enum
+import functools
 import importlib
+import importlib.util
 import logging
 import os
 import site
@@ -20,21 +22,26 @@ class Checkpoint(enum.IntEnum):
 SITE_PACKAGES = site.getsitepackages() + [
     site.getuserbase(),
     site.getusersitepackages(),
-    "/usr/lib/python",
 ]
 
-CACHEDIR = Path(".pytest_cache/forkserver")
+CACHEDIR = Path(".forkserver_cache")
 CACHEDIR.mkdir(parents=True, exist_ok=True)
 
 
 def write_modules_to_checkpoints() -> None:
     d = defaultdict(list)
-    for module in sys.modules:
+    f = CACHEDIR / "imported_modules"
+    if not f.exists():
+        return
+
+    for module in f.read_text().splitlines():
         cp = get_checkpoint(module)
         d[cp].append(module)
 
     for cp in d:
         get_file(cp).write_text("\n".join(d[cp]))
+
+    f.unlink(missing_ok=True)
 
 
 def load_modules_in_checkpoint(cp: Checkpoint) -> None:
@@ -76,16 +83,26 @@ checkpoints = [
 
 
 def _is_package(module: str) -> bool:
-    module = importlib.import_module(module)
+    prefix = module.split(".", 1)[0]
+    return _is_package_by_prefix(prefix)
 
-    # builtins dont have file?
-    f = getattr(module, "__file__", None)
-    if not f:
+
+@functools.cache
+def _is_package_by_prefix(prefix: str) -> bool:
+    if prefix in sys.stdlib_module_names:
         return True
-
-    f: str = os.path.abspath(module.__file__)
-    return any(f.startswith(sp) for sp in SITE_PACKAGES)
+    spec = importlib.util.find_spec(prefix)
+    if spec is None:
+        return True
+    if spec.origin is None:
+        return True
+    origin = os.path.abspath(spec.origin)
+    return any(origin.startswith(sp) for sp in SITE_PACKAGES)
 
 
 def _is_test_package(module: str) -> bool:
-    return True if "test" in module else False
+    return True if _is_main(module) and "test" in module else False
+
+
+def _is_main(module: str) -> bool:
+    return not _is_package(module)
